Fix ray tracing engine emission/incidence angles
* The NAIF DSK, Bullet abd Embree ray tracing engines were not computing ellipsoid and local (shape model) emission and incidence angles. This update makes them consistent in their behavior.

The bug is actually in ShapeModel::emissionAngle() and ShapeModel::incidenceAngle(). Neither explicitly computes a surface normal vector but tests to see if one is already computed and uses whatever happens to be there. There is no way to determine if the normal is computed from the ellipsoid or the shape model. Shape models can enforce this by reimplementing this method, but only Embree had this method until USGS removed emissionAngle(), thinking this would fix the problem – this actually created the ellipsoid emission angle consistency problem in phocube.

Signed-off-by: Kris J. Becker <kbecker@orex.lpl.arizona.edu>