Rather significant refactor of PSMRTS buffering
Integrating Bullet necessitated another look at PSMRTS’ data buffering system. This commit provides a heirarchy of buffer mapping of various structures to a fundamental data buffer handler named PsmrtsBufferData.

PsmrtsBufferData allocates a data buffer of bytes and provides a few indexing, access and management methods. The data is designed to be shared by default, but provides a deep copy method. It creates and manages slices and provides a mechanism for callers mapping a pointer to data in its own space.

PsmrtsStridingBuffer slices and dices a PsmrtsBufferData. This provides support for all data typing needs in PSMRTS including support for a variety of data management needs for diverse data formats and tracers.

PsmrtsBuffer specifies types using striding buffers. All indexing and operations are performed on arbitrary byte size striding sizes of a type T (which could itself be strided for arrays ot T).

PsmrtsVector3<T> provides a mapping to 3 element data vectors to type T. This uses PsmrtsBuffer<T> to create an efficent map for Eigen::Vectors. Efficient memory management is achieved using Eigen::Map<T> to provide read/write pointer referencing to a memory provided by our PSMRTS buffering system.

Some tests were impacted by this change as were much of the tracer and format systems. They were mostly fixed, but the Bullet-DSK float test is failing because of precision issues. This will be fixed in a n upcoming commit.

This refactor renders a few classes obsolete. The PsmrtsDataModel is essentialy replaced by PsmrtsVector3<T>. In fact, it is currently aliased in the code as PsmrtsVector3<T>. So the test_PsmrtsDataModel.cp needs to be renamed test_PsrmrtsVector3.cpp. This will occur in an upcoming commit.

 Also PsmrtsFormatModel is fully accomodated with tools provided in this refactor. The PsmrtsMeshMap now provides a solution for containing an mesh of any data type.

PsmrtsBulletMeshMap is much improved with this refactor and allows for locally maintained memory management and resuse of the PsmrtsBulletWorldModel tracer.