Develop/Implement performance tracking interface
This commit adds fundametal thread-safe time tracking and counters for tracers, format readers and data buffers and meshes.

Tracers track ray traces. There is a separate tracker for derived instances of PsmrtsTracerModels since it is highly likely trace counts will differ since they are shared.

Format reader load time can be tracked as well as the count of the number of meshes that were extracted from the format system.

Buffer and mesh life times are tracked, how many times it is cloned and its current number of shared instances of the buffer/mesh can be tracked at any time.