Major restructure of PSMRTS directory hierachy
In order to support export and (arguably) make maintenence easier, this is the proposed structure. The primary reason for moving all source code under a psmrts directory as it makes building in place and installing so much easier. This way we only need to add a $<BUILD_INTERFACE:${PSMRTS_PROJECT_DIR}> and $<INSTALL_INTERFACE:include> with the psmrts top level directory being the root install dir. The include interfaces are the same.

- Move ./capi, ./tracers ./shapes to ./pmsrts
- Move ./shapes/*/*Shape.hpp to ./shapes
- Move ./tracers/*/*.Tracer.hpp to ./tracers
- In each ./tracers/*/src to ./tracers/*/private
- In each ./shapes/*/*/src to ./shapes/*/private
- Create ./tools and ./scripts directories and moved some files there.

This should help make the system cleaner and easier to maintain.