Cleaned up a lot of unneccessary code
Lots of files were removed that were obsolete. Much of the functionality was redundent and just added to the maintanence load.

We are also building on the algorithm functionality.

This commit provides a more complete revision of the PsmrtsParameter class. This class will provide user/dev parameter support mainly limited to intrinsic types and arrays of intrinsics.