Update for ISIS cspice, fix code coverage
In the ISIS system, the find module only defines the include directory and library name so an additional option was added to the standard dependency macro.

Turns out adding the psmrts library broke code coverage. Needs to add the code coverage flags explicitly to the PSMRTS INTERFACE portion to pick up headers and prevent link errors.