FIx selection of package manager in build scripts
This commit provides support in the build scripts to select either VCPKG or conda as the dependency package manager. It also supports using a user- or system-based vcpkg install as well as create a local vcpkg installation in the ./psmrts directory. It will remain until the user manuallly deletes it.